refactor(page): extract compare mode into hooks/useCompareMode
app/page.tsx had grown to 622 lines with the compare-mode state,
effects and handlers inlined. Pulled them into a dedicated hook,
mirroring the shape of useShiftActions/useNoteActions — behavior
unchanged, verified with tsc/lint/build and a browser smoke test.
