+++ app/page.tsx
@@ -3,15 +3,14 @@
 import { useState, useEffect, Suspense, useCallback, useMemo } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useTranslations } from "next-intl";
-import { isSameDay, isSameMonth } from "date-fns";
-import { toast } from "sonner";
+import { isSameMonth } from "date-fns";
 import { ShiftWithCalendar } from "@/lib/types";
 import { CalendarNote } from "@/lib/db/schema";
 import { useCalendars } from "@/hooks/useCalendars";
 import { useShifts } from "@/hooks/useShifts";
 import { usePresets } from "@/hooks/usePresets";
 import { useNotes } from "@/hooks/useNotes";
-import { useCompareData } from "@/hooks/useCompareData";
+import { useCompareMode } from "@/hooks/useCompareMode";
 import { useViewSettings } from "@/hooks/useViewSettings";
 import { useQueryClient } from "@tanstack/react-query";
 import { queryKeys } from "@/lib/query-keys";
@@ -46,12 +45,6 @@ function toDayLayout(view: CalendarViewSettings): DayLayoutOptions {
     sortOrder: view.sortOrder,
     combinedSort: view.combinedSort,
   };
-}
-
-function toDate(date: Date | string): Date {
-  return typeof date === "string" && /^\d{4}-\d{2}-\d{2}$/.test(date)
-    ? parseLocalDate(date)
-    : new Date(date);
 }
 
 function HomeContent() {
@@ -95,9 +88,6 @@ function HomeContent() {
   const [selectedPresetId, setSelectedPresetId] = useState<string | undefined>();
   const [currentDate, setCurrentDate] = useState(new Date());
   const [selectedDate, setSelectedDate] = useState<Date | undefined>();
-  const [compareNoteCalendarId, setCompareNoteCalendarId] = useState<
-    string | undefined
-  >();
   const [editingShift, setEditingShift] = useState<ShiftWithCalendar | undefined>();
   const [daySheetOpen, setDaySheetOpen] = useState(false);
 
@@ -115,20 +105,32 @@ function HomeContent() {
     [selectedCalendar]
   );
 
-  const [isCompareMode, setIsCompareMode] = useState(false);
-  const [showCompareSelector, setShowCompareSelector] = useState(false);
-  const [selectedCompareIds, setSelectedCompareIds] = useState<string[]>([]);
-  // Selection before the picker opened, restored on cancel
-  const [compareSnapshot, setCompareSnapshot] = useState<string[]>([]);
-  const [compareTogglingDates, setCompareTogglingDates] = useState<
-    Map<string, Set<string>>
-  >(new Map());
-
   const queryClient = useQueryClient();
 
-  const compareData = useCompareData({
-    calendarIds: selectedCompareIds,
-    enabled: isCompareMode,
+  const {
+    isCompareMode,
+    setIsCompareMode,
+    showCompareSelector,
+    setShowCompareSelector,
+    selectedCompareIds,
+    setSelectedCompareIds,
+    compareSnapshot,
+    compareTogglingDates,
+    compareNoteCalendarId,
+    setCompareNoteCalendarId,
+    compareData,
+    handleToggleCompareCalendar,
+    handleExitCompare,
+    openComparePicker,
+    handleCompareDayClick,
+  } = useCompareMode({
+    calendars,
+    hasLoadedOnce,
+    selectedCalendar,
+    currentDate,
+    setCurrentDate,
+    selectDay,
+    selectedPresetId,
   });
 
   const {
@@ -192,48 +194,10 @@ function HomeContent() {
     dialogStates.setShowDayShiftsDialog(true);
   };
 
-  // Load compare mode from URL on initial load
-  useEffect(() => {
-    const compareParam = searchParams.get("compare");
-    if (compareParam && !isCompareMode) {
-      const calendarIds = compareParam.split(",").filter((id) => id.trim());
-      if (calendarIds.length >= 2 && calendarIds.length <= 3) {
-        const validIds = calendarIds.filter((id) =>
-          calendars.some((cal) => cal.id === id)
-        );
-        if (validIds.length >= 2) {
-          // Syncing from the URL (an external source), not from render state.
-          // eslint-disable-next-line react-hooks/set-state-in-effect
-          setSelectedCompareIds(validIds);
-          setIsCompareMode(true);
-        }
-      }
-    } else if (!compareParam && isCompareMode) {
-      setIsCompareMode(false);
-      setSelectedCompareIds([]);
-      setCompareTogglingDates(new Map());
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [searchParams, calendars]);
-
-  // The effect above only filters while entering compare mode. Access to one of
-  // the columns can be lost afterwards, and compare needs at least two.
-  useEffect(() => {
-    if (!isCompareMode || !hasLoadedOnce) return;
-    const stillAccessible = selectedCompareIds.filter((id) =>
-      calendars.some((cal) => cal.id === id)
-    );
-    if (stillAccessible.length === selectedCompareIds.length) return;
-    queueMicrotask(() => {
-      if (stillAccessible.length >= 2) {
-        setSelectedCompareIds(stillAccessible);
-      } else {
-        setIsCompareMode(false);
-        setSelectedCompareIds([]);
-        setCompareTogglingDates(new Map());
-      }
-    });
-  }, [isCompareMode, hasLoadedOnce, calendars, selectedCompareIds]);
+  const openCompareNotes = (calendarId: string, date: Date) => {
+    setCompareNoteCalendarId(calendarId);
+    openNotesForDay(compareData.notesMap.get(calendarId) || [], date);
+  };
 
   useEffect(() => {
     if (isCompareMode && selectedCompareIds.length >= 2) {
@@ -320,95 +284,6 @@ function HomeContent() {
     } else if (!desktop) {
       setDaySheetOpen(true);
     }
-  };
-
-  // Compare mode handlers
-  const handleToggleCompareCalendar = (calendarId: string) => {
-    setSelectedCompareIds((prev) =>
-      prev.includes(calendarId)
-        ? prev.filter((id) => id !== calendarId)
-        : [...prev, calendarId]
-    );
-  };
-
-  const handleExitCompare = () => {
-    setIsCompareMode(false);
-    setSelectedCompareIds([]);
-    setCompareTogglingDates(new Map());
-    router.replace(selectedCalendar ? `/?id=${selectedCalendar}` : `/`, {
-      scroll: false,
-    });
-  };
-
-  const updateCompareToggling = (calendarId: string, dateKey: string, add: boolean) => {
-    setCompareTogglingDates((prev) => {
-      const updated = new Map(prev);
-      const next = new Set(updated.get(calendarId) || []);
-      if (add) next.add(dateKey);
-      else next.delete(dateKey);
-      updated.set(calendarId, next);
-      return updated;
-    });
-  };
-
-  const openComparePicker = () => {
-    setCompareSnapshot(isCompareMode ? selectedCompareIds : []);
-    setShowCompareSelector(true);
-  };
-
-  const handleCompareDayClick = async (calendarId: string, date: Date | string) => {
-    const targetDate = toDate(date);
-    selectDay(targetDate);
-    if (!isSameMonth(targetDate, currentDate)) setCurrentDate(targetDate);
-    if (!selectedPresetId) return;
-
-    const calendarPresets = compareData.presetsMap.get(calendarId) || [];
-    const calendarShifts = compareData.shiftsMap.get(calendarId) || [];
-    const preset = calendarPresets.find((p) => p.id === selectedPresetId);
-    if (!preset) return;
-
-    const dateKey = formatDateToLocal(targetDate);
-    if (compareTogglingDates.get(calendarId)?.has(dateKey)) return;
-    updateCompareToggling(calendarId, dateKey, true);
-
-    try {
-      const existingShift = calendarShifts.find(
-        (shift) =>
-          shift.date &&
-          isSameDay(shift.date as Date, targetDate) &&
-          shift.title === preset.title &&
-          shift.startTime === preset.startTime &&
-          shift.endTime === preset.endTime
-      );
-
-      if (existingShift) {
-        await compareData.deleteShift({ calendarId, shiftId: existingShift.id });
-      } else {
-        await compareData.createShift({
-          calendarId,
-          formData: {
-            date: dateKey,
-            startTime: preset.startTime,
-            endTime: preset.endTime,
-            title: preset.title,
-            color: preset.color,
-            notes: preset.notes || "",
-            presetId: preset.id,
-            isAllDay: preset.isAllDay || false,
-          },
-        });
-      }
-    } catch (error) {
-      console.error("Failed to toggle shift:", error);
-      toast.error(t("common.error"));
-    } finally {
-      updateCompareToggling(calendarId, dateKey, false);
-    }
-  };
-
-  const openCompareNotes = (calendarId: string, date: Date) => {
-    setCompareNoteCalendarId(calendarId);
-    openNotesForDay(compareData.notesMap.get(calendarId) || [], date);
   };
 
   const dialogManager = (
